isdProject/requirements.txt
asgiref==3.5.2
cached-property==1.5.2
Django==4.0.4
django-cors-headers==3.12.0
django-url-filter @ git+https://github.com/bernd-wechner/django-url-filter@d0ae1ab23cf58bc800048eb21e2423c0e3efbda0
djangorestframework==3.13.1
enum-compat==0.0.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
